Fix, Fail-Proof
So, there was this tiny issue in just taking the diff between two 'sq' outputs. When jobs are running, their times change, making it a tossup as to whether or not the correct ID number was extracted from temp files.
I had this crazy idea that instead of parsing 'sq' outputs with temp files, we just take the stdout from 'sbatch' when a job is submitted and use that for the next job.
This has resulted in significant speedups and no more failures.